chore(state-nav): move to common directory
