+++ src/templates/state/index.js
@@ -1,7 +1,7 @@
 import React from 'react'
 import { graphql } from 'gatsby'
 import Layout from '~components/layout'
-import StateNavWrapper from '~components/pages/data/state-nav-wrapper'
+import StateNavWrapper from '~components/common/state-nav-wrapper'
 import StatePreamble from '~components/pages/state/preamble'
 import SummaryCharts from '~components/pages/data/summary-charts'
 import StateSummary from '~components/pages/data/summary'
